backend/requirements.txt
Django>=5.0.4
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
celery>=5.3.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
requests>=2.31.0